Fixed async problems with quiz and protectedroute.

+++ pc.quiz.client/src/components/Route/ProtectedRoute.tsx
@@ -1,8 +1,17 @@
 import { Navigate } from 'react-router-dom';
 import { useQuiz } from '../../utils/QuizContext';
+import { Box, CircularProgress } from '@mui/material';
 
 const ProtectedRoute = ({ element }: { element: JSX.Element }) => {
-    const { currentEntryId } = useQuiz();
+    const { currentEntryId, loading } = useQuiz();
+
+    if (loading) {
+        return (
+            <Box sx={{ display: 'flex', justifyContent: 'center', alignItems: 'center', height: '100vh' }}>
+                <CircularProgress />
+            </Box >
+        );
+    }
 
     if (!currentEntryId) {
         localStorage.clear();
